Support descriptions for commands, and improve command help output
This allows each command to have a description, and makes command
help show the description for the current command at the top, and
include the description for each subcommand when showing the
subcommand.  Command help output now uses the following format
(with sections omitted if they would be empty):

```
DESCRIPTION

Usage:
    BANNERS

Commands:
    COMMAND    DESCRIPTION

Post Commands:
    COMMAND    DESCRIPTION

Options:
    OPTION     DESCRIPTION

Post Options:
    OPTION     DESCRIPTION
```

This required restructuring how command help and option parsers work.
Previously, the option parser had a reference to the subcommand hash,
so it could include subcommand information. This removes that, and
makes changes to no longer use the option parser for banners.  Banners
are now stored as part of the command itself.  This is done to make
it easier to control command help output.

Other changes:

* Since the banner is no longer stored in the option parser,
  the SkipOptionParser class is no longer needed.  A :skip symbol is
  used instead.

* options and post_options in the DSL can no longer be called without
  a block.  Callers should switch to using banner and post_banner in
  cases where a block is not needed.  This avoids creating unnecessary
  option parsers in cases where there is just a banner and no options.

* process_command_failure no longer takes an option parser argument.

* The CommandFailure exception class now takes a command instead of
  an array of option parsers.

* Command help is precomputed and frozen when a command is frozen,
  making help for frozen commands very fast.

This is backwards incompatible, so the next release will include a
major bump.