Implement a plugin system
This changes Rodish to use a plugin system approach.  All parts of
Rodish can be configured via plugins.  To make the plugin system
viable, subclasses of DSL, Command, and OptionParser are created
per processor.  Plugins support the following methods to configure
these subclasses:

* ProcessorMethods
* DSLMethods
* CommandMethods
* OptionParserMethods

Plugin before_load/after_load methods are also supported.

By design, plugins are per processor, not per command.  Per-command
plugins would require each command be a subclass instead of an
instance, and I don't want that.

As a test, this moves Processor#usages to a plugin.  Other parts of
currently core Rodish will be moved to plugins.

As the plugin method is called in processor context and not command
context, you cannot call it inside the Rodish.processor block.
Support calling Rodish.processor without a block.  Allow calling
Processor#on method with a block and no command name, to support
configuring the root command.  Change recommended setup to be:

  Rodish.processor(self)

  # plugin :foo

  on do
    # configure root command
  end

This makes the API more Roda-like.